# spring-boot-bootcamp-resources-main/08. rest-api/contacts/starter/contacts/src/main/java/com/ltp/contacts/pojo/Contact.java
package com.ltp.contacts.pojo;

import javax.validation.Valid;
import javax.validation.constraints.NotBlank;
import java.util.UUID;


public class Contact {


    private String id;
    @NotBlank(message = "Name Cannot be Blank")
    private String name;
    @NotBlank(message = "Phone Number cannot be blank")
    private String phoneNumber;


//    public Contact(String id, String name, String phoneNumber) {
//        this.id = id;
//        this.name = name;
//        this.phoneNumber = phoneNumber;
//    }

    public Contact() {
        this.id = UUID.randomUUID().toString();
    }

    public String getName() {
        return this.name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getPhoneNumber() {
        return this.phoneNumber;
    }

    public void setPhoneNumber(String phoneNumber) {
        this.phoneNumber = phoneNumber;
    }

    public String getId() {
        return this.id;
    }

    public void setId(String id) {
        this.id = id;
    }

}